Week 4 - Write a Test (Programming Assignment)
In this exercise, you will be checking the accuracy of a string input to a given function against some conditions. You will be writing two functions. The first function will check if the length of the input string is within a specific limit of words and characters. The second function will check if the basic grammar of the string is well-defined.